requirements.txt
python-dotenv>=0.19.0
openai-whisper
sounddevice
numpy
openai
TTS
pydub
boto3==1.36.12
google-api-core==2.24.1
coqui-tts==0.25.3
google-cloud-texttospeech>=2.24.0
